added first onboarding page to app
